# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ResolveTitleTranslationsRequest": {
        "additionalProperties": false,
        "properties": {
          "content_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Content Ids",
            "type": "array"
          }
        },
        "required": [
          "content_ids"
        ],
        "title": "ResolveTitleTranslationsRequest",
        "type": "object"
      },
      "ResolveTranslationSegmentsRequest": {
        "additionalProperties": false,
        "properties": {
          "segments": {
            "items": {
              "$ref": "#/components/schemas/TranslationSegmentRequest"
            },
            "maxItems": 100,
            "minItems": 1,
            "title": "Segments",
            "type": "array"
          }
        },
        "required": [
          "segments"
        ],
        "title": "ResolveTranslationSegmentsRequest",
        "type": "object"
      },
...
      "TitleTranslationResponse": {
        "properties": {
          "content_id": {
            "title": "Content Id",
            "type": "string"
          },
          "engine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Id"
          },
          "engine_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Label"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Retryable"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "translated_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Title"
          },
          "translation_locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Locale"
          },
          "translation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Status"
          }
        },
        "required": [
          "content_id",
          "translated_title",
          "translation_locale",
          "translation_status",
          "engine_id",
          "engine_label",
          "error_code",
          "error_retryable",
          "retry_after_ms"
        ],
        "title": "TitleTranslationResponse",
        "type": "object"
      },
...
      "TranslationPurpose": {
        "description": "Product role of source text submitted for translation.",
        "enum": [
          "title",
          "excerpt",
          "body",
          "paragraph",
          "ranking_title",
          "ranking_description",
          "web_segment",
          "caption"
        ],
        "title": "TranslationPurpose",
        "type": "string"
      },
      "TranslationQuotaErrorDetail": {
        "additionalProperties": false,
        "description": "Safe, stable error detail returned for translation quota failures.",
        "properties": {
          "code": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "error_code": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Error Code",
            "type": "string"
          }
        },
        "required": [
          "code",
          "error_code"
        ],
        "title": "TranslationQuotaErrorDetail",
        "type": "object"
      },
      "TranslationQuotaErrorResponse": {
        "additionalProperties": false,
        "description": "FastAPI's HTTPException envelope for quota failures.",
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/TranslationQuotaErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "TranslationQuotaErrorResponse",
        "type": "object"
      },
      "TranslationSegmentRequest": {
        "additionalProperties": false,
        "properties": {
          "purpose": {
            "$ref": "#/components/schemas/TranslationPurpose"
          },
          "segment_id": {
            "maxLength": 160,
            "minLength": 1,
            "title": "Segment Id",
            "type": "string"
          },
          "text": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "segment_id",
          "text",
          "purpose"
        ],
        "title": "TranslationSegmentRequest",
        "type": "object"
      },
      "TranslationSegmentResponse": {
        "properties": {
          "cache_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Expires At"
          },
          "effective_engine_fingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Engine Fingerprint"
          },
          "engine_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Id"
          },
          "engine_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Engine Label"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_retryable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Retryable"
          },
          "purpose": {
            "$ref": "#/components/schemas/TranslationPurpose"
          },
          "retry_after_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After Ms"
          },
          "segment_id": {
            "title": "Segment Id",
            "type": "string"
          },
          "translated_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translated Text"
          },
          "translation_locale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Locale"
          },
          "translation_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Translation Status"
          }
        },
        "required": [
          "segment_id",
          "purpose",
          "translated_text",
          "translation_locale",
          "translation_status",
          "engine_id",
          "engine_label",
          "error_code",
          "error_retryable",
          "retry_after_ms"
        ],
        "title": "TranslationSegmentResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "ReaderServerToken": {
        "description": "Deployment connection token configured by the server administrator.",
        "in": "header",
        "name": "X-Reader-Server-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend for shared-source information aggregation.",
    "title": "Reader API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/translations/segments": {
      "post": {
        "description": "Resolve bounded UI segments for ranking, reader, WebView, and captions.\n\nWeb page text is user-scoped because a WebView may contain authenticated or\notherwise private content. Public ranking, article, and caption text may\nsafely reuse application-managed artifacts.",
        "operationId": "resolve_translation_segments_v1_translations_segments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveTranslationSegmentsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TranslationSegmentResponse"
                  },
                  "title": "Response Resolve Translation Segments V1 Translations Segments Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationQuotaErrorResponse"
                }
              }
            },
            "description": "Translation is disabled for this user."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationQuotaErrorResponse"
                }
              }
            },
            "description": "The authenticated request or translation miss budget is exhausted.",
            "headers": {
              "Retry-After": {
                "description": "Seconds until the fixed UTC quota window resets.",
                "schema": {
                  "minimum": 1.0,
                  "type": "integer"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationQuotaErrorResponse"
                }
              }
            },
            "description": "Translation storage or its durable quota store is unavailable."
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Translation Segments",
        "tags": [
          "translation"
        ]
      }
    },
    "/v1/translations/titles": {
      "post": {
        "description": "Resolve only loaded, accessible titles so clients can converge cheaply.",
        "operationId": "resolve_title_translations_v1_translations_titles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveTitleTranslationsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TitleTranslationResponse"
                  },
                  "title": "Response Resolve Title Translations V1 Translations Titles Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ReaderServerToken": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Title Translations",
        "tags": [
          "translation"
        ]
      }
    },
...
  }
}